Replaced HSQLDB for H2. Much better performance (less CPU consumption and faster results on larger sets, will try full text search support next)

+++ gui/com/limegroup/gnutella/gui/options/panes/SmartSearchDBPaneItem.java
@@ -24,8 +24,8 @@ public final class SmartSearchDBPaneItem extends AbstractPaneItem {
 	private JLabel _numTorrentsLabel;
 	private JLabel _numFilesLabel;
 	
-	private int _numTorrents = 0;
-	private int _numFiles = 0;
+	private long _numTorrents = 0;
+	private long _numFiles = 0;
  
 
     /**
